requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
fastapi>=0.128.0
uvicorn>=0.40.0
